requirements.txt
flet
eyed3
mutagen
pillow
tempfile
os
io